refactor(user-management): replace IUserSearchRepository with IElasticSearchService
Move user search functionality from domain repository to application service layer to better align with service boundaries. Removed unused admin repository interfaces and implementations.
