+++ services/user-admin-service/src/application/interface/events/IEventBus.ts
@@ -1,0 +1,7 @@
+import { IEventHandler } from "./handlers/IEventHandler";
+import { DomainEvent } from "../../../domain/events/DomainEvent";
+
+export interface IEventBus {
+    register<T extends DomainEvent>(eventType: string, handler: IEventHandler<T>): void;
+    publish(event: DomainEvent): Promise<void>;
+}
