+++ services/user-admin-service/src/application/interface/http/IElasticSearchService.ts
@@ -1,0 +1,3 @@
+export interface IElasticSearchService {
+    searchUserIds(query: string): Promise<string[]>;
+}
