+++ services/user-admin-service/src/application/usecases/admin/dashboard/GetPlatformRevenueStatsUseCase.ts
@@ -1,6 +1,6 @@
 import { injectable, inject } from 'inversify';
 import { TYPES } from '../../../../infrastructure/inversify/types';
-import { IGetPlatformRevenueStatsUseCase, PlatformRevenueStats } from '../../../interface/usecase/admin/IGetPlatformRevenueStatsUseCase';
+import { IGetPlatformRevenueStatsUseCase, PlatformRevenueStats } from '../../../interface/usecases/admin/IGetPlatformRevenueStatsUseCase';
 import { GetPlatformRevenueStatsDTO } from '../../../interface/dtos/admin/GetPlatformRevenueStatsDTO';
 import { IWalletService } from '../../../interface/http/IWalletService';
