--- services/user-admin-service/src/domain/repositories/admin/IAdminRepository.ts
@@ -1,5 +1,0 @@
-import { IBaseRepository } from '../IBaseRepository';
-import { SafeUser, User } from '../../entities/User';
-
-export interface IAdminRepositories extends IBaseRepository<User, SafeUser> {
-}
