--- services/user-admin-service/src/domain/repositories/user/IUserSearchRepository.ts
@@ -1,3 +1,0 @@
-export interface IUserSearchRepository {
-  searchUserIds(query: string): Promise<string[]>;
-}
