+++ services/user-admin-service/src/infrastructure/http/ElasticSearchService.ts
@@ -1,10 +1,10 @@
 import axios from 'axios';
 import { injectable } from 'inversify';
-import { IUserSearchRepository } from '../../../domain/repositories/user/IUserSearchRepository';
-import { ROUTES } from '../../../constants/routes';
+import { ROUTES } from '../../constants/routes';
+import { IElasticSearchService } from '../../application/interface/http/IElasticSearchService';
 
 @injectable()
-export class ElasticUserSearchRepositoryImpl implements IUserSearchRepository {
+export class ElasticSearchService implements IElasticSearchService {
   async searchUserIds(query: string): Promise<string[]> {
     const response = await axios.get(
       `http://api-gateway:4000${ROUTES.EXTERNAL.ELASTIC_SEARCH}`,
